+++ client/public/components/dropdown/dropdown.js
@@ -1,4 +1,4 @@
-import Component from '../core/component.js';
+import Component from '../../core/component.js';
 
 import './dropdown.scss';
