+++ client/public/components/list/list.scss
@@ -1,4 +1,4 @@
-@import '../assets/woowa-market-icons.css';
+@import '../../assets/woowa-market-icons.css';
 
 .product-list {
   width: 100%;
