+++ client/public/components/modal/modal.js
@@ -1,5 +1,5 @@
-import Component from '../core/component.js';
-import _ from '../utils/dom.js';
+import Component from '../../core/component.js';
+import _ from '../../utils/dom.js';
 
 import './modal.scss';
