+++ client/public/components/modal/modal.scss
