--- client/public/components/slider.scss
@@ -1,73 +1,0 @@
-.slider__container {
-  width: 100%;
-	height: 650px;
-  background: var(--white-color);
-
-	&.open {
-		animation: open 0.2s forwards;
-	}
-	&.close {
-		animation: close 0.2s forwards;
-	}
-}
-
-@keyframes close {
-	0% {
-		transform: translateX(0%);
-	}
-	100% {
-		transform: translateX(100%);
-	}
-}
-
-@keyframes open {
-	0% {
-		transform: translateX(100%);
-	}
-	100% {
-		transform: translateX(0%);
-	}
-}
-
-@keyframes easeOutBounce {
-	0% {
-		transform: translateX(0%);
-	}
-
-	12% {
-		transform: translateX(-10.89%);
-	}
-
-	24% {
-		transform: translateX(-43.56%);
-	}
-
-	36% {
-		transform: translateX(-98.01%);
-	}
-
-	54% {
-		transform: translateX(-90.02%);
-	}
-
-	74% {
-		transform: translateX(-98.37%);
-	}
-
-	82% {
-		transform: translateX(-96.75%);
-	}
-
-	92% {
-		transform: translateX(-99.34%);
-	}
-
-	96% {
-		transform: translateX(-98.76%);
-	}
-
-	100% {
-		transform: translateX(-100%);
-	}
-
-}
