+++ client/public/components/slider/slider.scss
@@ -1,0 +1,65 @@
+.slider__container {
+  width: 100%;
+	height: 650px;
+  background: var(--white-color);
+	display: flex;
+	flex-direction: column;
+
+	&.open {
+		animation: easeOutBounce 0.5s forwards;
+	}
+	&.close {
+		animation: close 0.2s forwards;
+	}
+}
+
+@keyframes close {
+	0% {
+		transform: translateX(0%);
+	}
+	100% {
+		transform: translateX(-100%);
+	}
+}
+
+@keyframes open {
+	0% {
+		transform: translateX(-100%);
+	}
+	100% {
+		transform: translateX(0%);
+	}
+}
+
+@keyframes easeOutBounce {
+	0% {
+		transform: translateX(-100%);
+	}
+	12% {
+		transform: translateX(-89.11%);
+	}
+	24% {
+		transform: translateX(-56.44%);
+	}
+	36% {
+		transform: translateX(-1.99%);
+	}
+	54% {
+		transform: translateX(-25%);
+	}
+	74% {
+		transform: translateX(-1.63%);
+	}
+	82% {
+		transform: translateX(-6.25%);
+	}
+	92% {
+		transform: translateX(-0.7%);
+	}
+	96% {
+		transform: translateX(-1.5%);
+	}
+	100% {
+		transform: translateX(0%);
+	}
+}
